combined *.d.ts files and *-tests.ts
