--- webgme/metadata.json
@@ -1,26 +1,0 @@
-{
-    "id": "SamplePlugin",
-    "name": "A Sample Metadata for the Sample Plugin",
-    "version": "0.1.0",
-    "description": "This is a browser based serializer plugin",
-    "icon": {
-        "class": "glyphicon glyphicon-cog",
-        "src": ""
-    },
-    "disableServerSideExecution": true,
-    "disableBrowserSideExecution": false,
-    "writeAccessRequired": true,
-    "configStructure": [
-        {
-            "name": "aParameterName",
-            "displayName": "A Parameter Name",
-            "description": "a parameter for the sample plugin",
-            "value": "default-value",
-            "valueType": "string",
-            "valueItems": [
-                "default-value",
-                "some-value"
-            ]
-        }
-    ]
-}
