--- webgme/webgme-blob.d.ts
@@ -1,15 +1,0 @@
-
-
-declare namespace Blob {
-
-    export type ObjectBlob = string;
-
-    export interface BlobMetadata {
-        name: string;
-        size: number;
-        mime: string;
-        context: Core.DataObject;
-        contentType: string;
-    }
-    export interface Util { }
-}
