--- webgme/webgme-client-tests.ts
@@ -1,27 +1,0 @@
-
-/**
- * https://github.com/webgme/webgme/wiki/GME-Blob-Storage-API
- * 
- * File-like objects/artifacts 
- * (which are neither a meta archetype model, nor an instance model) 
- * are stored separately from the WebGME meta-models and models. 
- * An example of such an artifact would be a resource 
- * file that is associated with a model 
- * (e.g., data for an instance model, or a generated artifact from analyzing a model).
- * 
- * One reason for treating these objects differently is that they do not conform to the data model, 
- * and they might not be well-suited for storage in a database 
- * (the Blob is suited to handle binary objects of any size and structure).
- */
-
-/**
- * https://github.com/webgme/webgme/wiki/GME-Blob-Storage-API#usage
- */
-function test_client_creating_an_instance() {
-    // let client = new GME.classes.Client(GME.gmeConfig);
-}
-
-/**
- * Add the remaining tests.
- */
-
